requirements.txt
google-genai
python-dotenv
requests
extract-msg
pymupdf
python-docx
beautifulsoup4
pandas
oracledb
unidecode